Small Fixes
Made armor unable to be damaged (by making it repair itself when damaged). Toggleable, however even if the head, legs, and feet pieces break, the chest piece is still unbreakable since the damage values are used for the air system.

Also added a revision to the Persian translation by Adoosth#2266 on Discord